Keep base image folders permissions when unpacking
We have to chmod it because creat and mkdir syscalls take the umask into
account when applying the permissions. This means that only permissions
less permissive than the umask can be applied to files and directories.
By calling chmod we explicitly apply the permissions without being
subject to the umask.

[fixes #133921963]

Signed-off-by: Tiago Scolari <tscolari@pivotal.io>