Image creation from volume ids rather than path
This is to support multi-layer images in the overlay driver.
In BTRFS we only care about the top layer since the nature of snapshots
means that the whole layer stack is visible through the top layer.
Conversely in overlay we need to mount an ordered stack of layers in
order to create an image. By passing the stack of volume ids to the
image creation functions in our driver, we can support this without
needing to store additional layer ordering metadata.

[#139825877]

Signed-off-by: Tiago Scolari <tscolari@pivotal.io>