Migrate all tests to use Runner instead of exec.Command
We need to do this because it's the Runner that's responsible for
controlling whether grootfs runs as root or non-root in the tests.

We've introduced a new StartCommand function to the Runner which can be
used to kick off grootfs without waiting for the command to complete.
Some tests just need the session rather than the final output.

[finishes #142687079]

Signed-off-by: Tiago Scolari <tscolari@pivotal.io>