Change chainID generation format for tar rootfs
The old format was sha256(rootfsPathname)-rootfsMtime. This was a little
weird, as the volume IDs of imported tars were SHAs hyphenated with
timestamps instead of SHAs *of* the pathname
concatenated/hyphenated/whatever with the mtime. This was harmless
however.

OCI images must contain digest identifiers, which wouldn't technically
be valid if they were SHAs hyphenated with timestamps. So, Guardian
generates identifiers for tar layers of OCI images using the format
sha256(rootfsPathname-rootfsMtime), and passes this OCI image manifest
to Groot.

When oci-phase-1 is enabled on Cloud Foundry, the staging container of a
buildpack app is based on a tar rootfs, and the running container is
based on an OCI image. The bottom layer of this OCI image happens to be
based on exactly the same tar, but Groot wasn't recognising them as
identical volumes because of the different identifier format.

Since Groot is configured to garbage collect before creating new images,
the imported staging rootfs was often deleted before an app was run.
This caused unnecessary delays deleting and re-importing the tar rootfs.

By using the same format, we can avoid this situation while oci-phase-1
is enabled. In the long run, OCI images will be used for everything
anyway.

Upgrading to this revision of groot will cause the tar rootfs to be
re-imported once per cell, and the cache entry with the old format will
eventually be garbage collected.

[#154844133]

Signed-off-by: Craig Furman <cfurman@pivotal.io>