Handle basedir in new userns when rootless
Basdir handling involves chowning the base directory and chowning is
only permitted to root. In order to make chowning work when
rootless, this commit would make sure that we reexec into a new user
namespace (so that we become root) prior any chowning

In cases when we want the reexecer to reexec into a new user namespace
it would perform double reexec. For some unknown reason if we rexeced
just once the new process has no capabilities and chown would fail with
permission denied error. Reexecing again magically provides the needed
capabilities and chown works fine. Go figure... ¯\_(ツ)_/¯

This huge commit involves:
* Introducing the `sandbox.Reexecer` utility that allows reexecing
arbitrary action in a sandboxed environment. Depending on the
`RexecSpec` provided to the `Reexec` method, the reexecer could
optionally rexec into a new usernamespace and/or chroot into a
predefined directory.
* Replacing all reexecs in the code with `sandbox.Reexecer`
* The decision whether to reexec into new user namespace is pulled up
into the commands
* Extracting standalone `BasedirHandler` utility from the `base_image_puller`
* Extracting `IDTranslator` from the tar unpacker
* Extracting `WhiteoutHandler` from the tar unpacker
* Updating the `cloudfoundry.experimental.image.base-directory` test OCI
image to contain a layer with a basedir

A big thank you goes to all commit authors who endured all the pain to
refactor all the reexec horror and fix all the unit tests.

[#162679612]

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Claudia Beresford <cberesford@pivotal.io>
Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>