Remove mountpoint check from DeInitFilesystem
This check is racey. It might be responsible for a flake we are
observing. Instead do an unconditional unmount and ingnore ENOENT and
EINVAL. The code is more robust this way and the potential race is
removed

Skip generate-volume-size-metadata test in rootless. This command is
only used in the `overlay-xfs-setup` scrip in garden-runc-release, which
is only supposed to run as root

[#164301588]

Co-authored-by: Giuseppe Capizzi <gcapizzi@pivotal.io>