Pin golang.org/x/exp to version v0.0.0-20230724220655-d98519c11495
We're pinning this repo to maintain compatibility with github.com/containers/image.

The containers/image module depends on the old interface for the 'slices' package. In older
versions of the 'slices' package, comparator functions return a bool; in a recent change from
July 25, the 'slices' package now wants comparators that return an int.

We expect that the containers/image repo will eventually be updated to use the new interface, at
which point, we can unpin this dependency.

Signed-off-by: Marc Paquette <mpaquette@vmware.com>